// src/models/ui/ComputeNodesUiModel.ts
export interface ComputeNodesData {
    name: string;
    ipAddressInput: string;
    framesFolderPath: string;
    membername: string;
    password: string;
    max_no_of_streams: number;
    toast_message: string;
    host_name: string;
    gpu_count: number;
    config_details: string;
    // createdBy: number;
}